Fixing publish entries type for graphql request
